+++ app/habits/page.tsx
@@ -3,9 +3,7 @@ import HabitList from '@/components/HabitList'
 
 export default function HabitsPage() {
   return (
-    <Layout>
-      <HabitList />
-    </Layout>
+    <HabitList />
   )
 }