+++ app/wishlist/page.tsx
@@ -3,9 +3,7 @@ import WishlistManager from '@/components/WishlistManager'
 
 export default function WishlistPage() {
   return (
-    <Layout>
-      <WishlistManager />
-    </Layout>
+    <WishlistManager />
   )
 }